+++ projects/_01_ingest/capacity_tracker/tests/fargate/test_clean_capacity_tracker_care_home_data.py
@@ -33,3 +33,4 @@ class TestMain:
             CTCHClean.ct_care_home_total_employed
             in returned_lf.collect_schema().names()
         )
+        assert returned_lf.collect()[CTCHClean.hours_agency][0] == 12.5